feat: add comment system with repository, service, router, and tests

+++ app/services/service_post.py
@@ -19,21 +19,23 @@ class PostService:
     async def get_all_posts(self,
                             user: User,
                             session: AsyncSession):
-        return await PostRepository().get_all_posts_of_the_user(user.id, session)
+        return await PostRepository().get_all(session)
 
     async def get_post_by_id(self,
                              post_id: int,
                              user: User,
                              session: AsyncSession):
-        return await PostRepository().get_by_id(post_id, user.id, session)
+        return await PostRepository().get_by_id(post_id, session)
 
     async def update_post(self,
                           post_id: int,
                           new_post: scheme_post.PostUpdate,
                           user: User,
                           session: AsyncSession):
-        result = await PostRepository().get_by_id(post_id, user.id, session)
-        if result:
+        result = await PostRepository().get_by_id(post_id, session)
+        if result is None:
+            return None
+        if result and result.owner_id == user.id:
             return await PostRepository().update(post_id,  session, new_post)
         return result
 
@@ -41,14 +43,16 @@ class PostService:
                           post_id: int,
                           user: User,
                           session: AsyncSession):
-        result = await PostRepository().get_by_id(post_id, user.id, session)
-        if result:
+        result = await PostRepository().get_by_id(post_id, session)
+        if result is None:
+            return None
+        if result and result.owner_id == user.id:
             return await PostRepository().delete(post_id,  session)
-        return result
+        return False
 
     async def search_post(self,
                           user: User,
                           session: AsyncSession,
                           title: Optional[str] = None,
                           is_published: Optional[bool] = None):
-        return await PostRepository().search(user.id, session, title=title, is_published=is_published)
+        return await PostRepository().search(session, title=title, is_published=is_published)
